requirements.txt
streamlit
firebase-admin
torch
torchvision
Pillow
numpy
